fixed update reducer to use Update<T>
